+++ src/js/components/footer.js
@@ -1,0 +1,14 @@
+import React, { Fragment } from 'react'
+import PropTypes from 'prop-types'
+
+import '../../scss/footer.scss'
+
+const Footer = () => (
+  <footer className={'home-page-footer'}>
+    <div className="home-page-footer-wrapper">
+      <p>Built with ❤ by @willmendesneto</p>
+    </div>
+  </footer>
+)
+
+export default Footer