+++ src/js/components/home-page.js
@@ -1,18 +1,24 @@
-import React from 'react';
+import React, { Fragment } from 'react'
 import { HowToStart } from 'generator-poi-boilerplate-demo'
 
 import '../../scss/styles.scss'
 
-// This component is generated automatically
-// As your first step, please remove `HowToStart` component reference
+import Header from './header'
+import Footer from './footer'
+import FirstSection from './first-section'
 
 const HomePage = ({ type }) => (
-  <section className={`home-page-wrapper is-${type}`}>
-    <h1 className="home-page-heading">
-      <p>This is your `<code className="home-page-code">home-page</code>` component</p>
-    </h1>
-    <HowToStart />
-  </section>
+  <Fragment>
+    <Header />
+    <FirstSection />
+    <section className={`home-page-wrapper is-${type}`}>
+      <h1 className="home-page-heading">
+        <p>This is your `<code className="home-page-code">home-page</code>` component</p>
+      </h1>
+      <HowToStart />
+    </section>
+    <Footer />
+  </Fragment>
 )
 
 export default HomePage